+++ tpl/data/resources.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package tplimpl
+package data
 
 import (
 	"bytes"
@@ -176,25 +176,25 @@ func resGetLocal(url string, fs afero.Fs, cfg config.Provider) ([]byte, error) {
 }
 
 // resGetResource loads the content of a local or remote file
-func (t *templateFuncster) resGetResource(url string) ([]byte, error) {
+func (ns *Namespace) resGetResource(url string) ([]byte, error) {
 	if url == "" {
 		return nil, nil
 	}
 	if strings.Contains(url, "://") {
-		return resGetRemote(url, t.Fs.Source, t.Cfg, http.DefaultClient)
+		return resGetRemote(url, ns.deps.Fs.Source, ns.deps.Cfg, http.DefaultClient)
 	}
-	return resGetLocal(url, t.Fs.Source, t.Cfg)
+	return resGetLocal(url, ns.deps.Fs.Source, ns.deps.Cfg)
 }
 
-// getJSON expects one or n-parts of a URL to a resource which can either be a local or a remote one.
+// GetJSON expects one or n-parts of a URL to a resource which can either be a local or a remote one.
 // If you provide multiple parts they will be joined together to the final URL.
 // GetJSON returns nil or parsed JSON to use in a short code.
-func (t *templateFuncster) getJSON(urlParts ...string) interface{} {
+func (ns *Namespace) GetJSON(urlParts ...string) interface{} {
 	var v interface{}
 	url := strings.Join(urlParts, "")
 
 	for i := 0; i <= resRetries; i++ {
-		c, err := t.resGetResource(url)
+		c, err := ns.resGetResource(url)
 		if err != nil {
 			jww.ERROR.Printf("Failed to get json resource %s with error message %s", url, err)
 			return nil
@@ -205,7 +205,7 @@ func (t *templateFuncster) getJSON(urlParts ...string) interface{} {
 			jww.ERROR.Printf("Cannot read json from resource %s with error message %s", url, err)
 			jww.ERROR.Printf("Retry #%d for %s and sleeping for %s", i, url, resSleep)
 			time.Sleep(resSleep)
-			resDeleteCache(url, t.Fs.Source, t.Cfg)
+			resDeleteCache(url, ns.deps.Fs.Source, ns.deps.Cfg)
 			continue
 		}
 		break
@@ -226,23 +226,23 @@ func parseCSV(c []byte, sep string) ([][]string, error) {
 	return r.ReadAll()
 }
 
-// getCSV expects a data separator and one or n-parts of a URL to a resource which
+// GetCSV expects a data separator and one or n-parts of a URL to a resource which
 // can either be a local or a remote one.
 // The data separator can be a comma, semi-colon, pipe, etc, but only one character.
 // If you provide multiple parts for the URL they will be joined together to the final URL.
 // GetCSV returns nil or a slice slice to use in a short code.
-func (t *templateFuncster) getCSV(sep string, urlParts ...string) [][]string {
+func (ns *Namespace) GetCSV(sep string, urlParts ...string) [][]string {
 	var d [][]string
 	url := strings.Join(urlParts, "")
 
 	var clearCacheSleep = func(i int, u string) {
 		jww.ERROR.Printf("Retry #%d for %s and sleeping for %s", i, url, resSleep)
 		time.Sleep(resSleep)
-		resDeleteCache(url, t.Fs.Source, t.Cfg)
+		resDeleteCache(url, ns.deps.Fs.Source, ns.deps.Cfg)
 	}
 
 	for i := 0; i <= resRetries; i++ {
-		c, err := t.resGetResource(url)
+		c, err := ns.resGetResource(url)
 
 		if err == nil && !bytes.Contains(c, []byte(sep)) {
 			err = errors.New("Cannot find separator " + sep + " in CSV.")
